Priority description lock

+++ src/features/tasks/server/route.ts
@@ -162,7 +162,17 @@ app.post(
         const user = c.get("user");
         const databases = c.get("databases");
 
-        const {name, status, workspaceId, projectId, dueDate, assigneeId, priority} = c.req.valid("json");
+        const {
+            name,
+            status,
+            workspaceId,
+            projectId,
+            dueDate,
+            assigneeId,
+            priority,
+            description,
+            locked
+        } = c.req.valid("json");
 
         const member = await getMember({
             databases,
@@ -195,6 +205,8 @@ app.post(
             assigneeId,
             position: newPosition,
             priority,
+            description,
+            locked,
         });
 
         return c.json({data: task});
@@ -208,7 +220,7 @@ app.patch(
     async (c) => {
         const user = c.get("user");
         const databases = c.get("databases");
-        const {name, status, dueDate, assigneeId, taskId, priority} = c.req.valid("json");
+        const {name, status, dueDate, assigneeId, taskId, priority, description, locked} = c.req.valid("json");
         let workspaceId: string = "";
 
         //get workspaceId from task
@@ -232,19 +244,32 @@ app.patch(
             return c.json({error: "Task not found"}, 404);
         }
 
-        //check if task is locked
-        // const task = await databases.getDocument<Task>(DATABASE_ID, TASKS_ID, taskId);
-        // if (task.locked && member.role === MemberRole.MEMBER) { //as string here?????
-        //     return c.json({error: "Task is locked"}, 403);
-        // }
-
         // construct update task object
         const updateData: Partial<Task> = {};
         if (name !== undefined) updateData.name = name;
         if (status !== undefined) updateData.status = status;
         if (dueDate !== undefined) updateData.dueDate = dueDate;
         if (assigneeId !== undefined) updateData.assigneeId = assigneeId;
-        if (priority !== undefined) updateData.priority = priority;
+        if (priority !== undefined) { // @ts-ignore
+            updateData.priority = priority;
+        }
+        if (description !== undefined) updateData.description = description;
+        if (locked !== undefined) updateData.locked = locked;
+
+
+        //check if task is locked
+        const task = await databases.getDocument<Task>(DATABASE_ID, TASKS_ID, taskId);
+        if (task.locked && member.role === MemberRole.MEMBER) {
+            return c.json({error: "Task is locked"}, 403);
+        } else if (task.locked && member.role === MemberRole.ADMIN) {
+            if (updateData.locked === undefined || updateData.locked === true) {
+                return c.json({error: "Task is locked, please unlock before making changes"}, 403);
+            }
+        }
+
+        if (updateData.locked !== undefined && member.role === MemberRole.MEMBER) {
+            return c.json({error: "Only admin can change task lock status"}, 403);
+        }
 
         //get userId by assigneeId
         let notificationUserId: string = "";
@@ -285,9 +310,10 @@ app.delete(
             workspaceId = task.workspaceId;
         });
 
+        let member;
         // check if user is member of workspace
         if (workspaceId) {
-            const member = await getMember({
+            member = await getMember({
                 databases,
                 workspaceId,
                 userId: user.$id,
@@ -296,6 +322,14 @@ app.delete(
             if (!member) {
                 return c.json({error: "Unauthorized"}, 401);
             }
+        } else {
+            return c.json({error: "Task not found"}, 404);
+        }
+
+        //check if task is locked
+        const task = await databases.getDocument<Task>(DATABASE_ID, TASKS_ID, taskId);
+        if (task.locked) {
+            return c.json({error: "Task is locked"}, 403);
         }
 
         await databases.deleteDocument(DATABASE_ID, TASKS_ID, taskId);
